add lex
