+++ markdown/src/defaults.js
@@ -4,10 +4,12 @@ function getDefaults() {
     breaks: false,
     gfm: true,
     headerIds: true,
+    mangle: true,
+
     headerPrefix: '',
     highlight: null,
     langPrefix: 'language-',
-    mangle: true,
+
     pedantic: false,
     renderer: null,
     sanitize: false,
