--- t.md
@@ -1,20 +1,0 @@
-**10
-11**
-
-** 12
-13**
-
-**14
-15 **
-
-**16
-17
-**
-
-** 18 **
-
-** 19**
-
-**20 **
-
-**21**
